Provide IlValue as condition for the Switch IlBuilder service
The existing Switch and  TableSwitch (and their overloads) services
could only be used if the condition to switch upon was a named
variable. It was not possible to provide an IlValue to switch on,
which limited the scenarios where we could use the Switch service.

This changeset adds 2 overloads of the Switch service to use IlValue
as the condition to switch on, while still ensuring that the IlValue
provided is of Int32 data type.

Signed-off-by: Nazim Uddin Bhuiyan <nazimudd@ualberta.ca>