Disable code cache reclamation if perfTool option is enabled
When the perfTool option is enabled the user is likely profiling JIT
generated code which means we would want to disable code cache
reclamation to avoid the profiling tool from mapping ticks to
overlapping JIT methods. We spit out a warning message to stderr so it
is obvious to the user that this is happening.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>