Reduce counts when running without SCC and many CPUs are available
When the shared class cache feature is not used, the default invocation counts
are count=3000,bcount=3000. They are rather large to allow for enough interpreter
profiling information to be collected. However, this affects the rampup of
Java applications because the transition from interpreter to jitted code is delayed.
Recent experiments have shown that it is possible to reduce the counts from
(3000,3000) to (2000,1000) without affecting the long term throughput.
Because lower invocation counts means more compilations, we can afford to lower the
counts only when we have enough computing power. In this PR the default
invocation counts will be lowered to (2000,1000) only when there are 4 CPUs
or more available.

Signed-off-by: Marius Pirvu <mpirvu@ca.ibm.com>