Fix segfault in omrvmemtest
As part ofr cleanup testOverlappingSegments deallocates all its memory
descriptrors. However if one of the allocations in the first pass fails,
the remaining portion of the array of descriptors remains uninitialized.
Attempting to later deallocate them causes a segfault. To solve the
issue, all descriptors are initialized to 0, and empty descriptors are
skipped during deallocation cleanup.

Signed-off-by: Devin Nakamura <devinn@ca.ibm.com>